requirements.txt
aiomysql==0.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
